Refactor : gamestart 계층 수정
